requirements/prod.txt
argparse==1.4.0
asgiref==3.8.1
Django==4.2.16
Pillow==11.3.0
sorl-thumbnail==12.11.0
django-cleanup==9.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
ffmpeg-python==0.2.0